+++ apps/api/migrations/0001_initial.py
@@ -2,6 +2,8 @@
 from __future__ import unicode_literals
 
 from django.db import migrations, models
+import api.validators
+
 
 class Migration(migrations.Migration):
 
@@ -12,31 +14,15 @@ class Migration(migrations.Migration):
         migrations.CreateModel(
             name='Log',
             fields=[
-                ('id', models.AutoField(serialize=False, auto_created=True, verbose_name='ID', primary_key=True)),
+                ('id', models.AutoField(verbose_name='ID', serialize=False, primary_key=True, auto_created=True)),
                 ('received', models.DateTimeField(auto_now_add=True)),
-                ('finished', models.DateTimeField(null=True, blank=True)),
-                ('history', models.TextField(default='', blank=True)),
-                ('nfc_id', models.CharField(max_length=255)),
-                ('station_id', models.CharField(max_length=20)),
+                ('finished', models.DateTimeField(blank=True, null=True)),
+                ('history', models.TextField(blank=True, default='')),
+                ('backpack_url', models.CharField(max_length=255, validators=[api.validators.is_url])),
+                ('station_id', models.PositiveIntegerField()),
             ],
             options={
                 'ordering': ['-received'],
-            },
-        ),
-        migrations.CreateModel(
-            name='NFCRegister',
-            fields=[
-                ('id', models.AutoField(serialize=False, auto_created=True, verbose_name='ID', primary_key=True)),
-                ('received', models.DateTimeField(auto_now_add=True)),
-                ('finished', models.DateTimeField(null=True, blank=True)),
-                ('history', models.TextField(default='', blank=True)),
-                ('created', models.DateTimeField(auto_now_add=True)),
-                ('nfc_id', models.CharField(unique=True, max_length=255)),
-                ('backpack_id', models.CharField(unique=True, max_length=6)),
-                ('language_id', models.IntegerField()),
-            ],
-            options={
-                'abstract': False,
             },
         ),
     ]