+++ apps/api/validators.py
@@ -1,0 +1,8 @@
+from .helpers import extract_backpack_id
+
+
+def is_url(value):
+    """
+    Check if given value has a backpack_id
+    """
+    extract_backpack_id(value)