+++ apps/base/config/common.py
@@ -171,4 +171,3 @@ class Common(Configuration):
         environ_prefix='',
         default=5
     )
-